requirements.txt
flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
numpy==1.24.3
beautifulsoup4==4.12.2
lxml==4.9.3
streamlit==1.28.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
openai==0.28.1
langchain==0.0.335
pydantic==1.10.8
pyPDF2==3.0.1
matplotlib==3.7.2